bridge: Add support of `IFLA_BR_MCAST_QUERIER_STATE`
Using `InfoBridge::MulticastQuerierState(Vec<BridgeQuerierState>)` for
`IFLA_BR_MCAST_QUERIER_STATE`.

Also isolated the `InfoBridge` related code to dedicate
`src/rtnl/link/nlas/bridge.rs` file.

Unit test case included using real netlink capture data.
